requirements.txt
requests>=2.28.0
colorama>=0.4.6
tqdm>=4.65.0
dnspython>=2.3.0
beautifulsoup4>=4.11.0
scapy>=2.5.0
